calling the lists from an external file
