+++ src/components/atoms/FilttersPost/functions.js
@@ -1,0 +1,16 @@
+const listData = [
+  {label: 'Todas', value: 1},
+  {label: 'Ayer', value: 2},
+  {label: 'Esta Semana', value: 3},
+];
+const listData2 = [
+  {label: 'Publicaciones vistas', value: 4},
+  {label: 'Publicaciones que te han gustado', value: 5},
+];
+
+const Lists = {
+  listData,
+  listData2,
+};
+
+export default Lists;
